+++ backend/src/routers/recipe_notes.py
@@ -1,29 +1,25 @@
-from fastapi import APIRouter, Depends, HTTPException
-from database import get_db
-from sqlalchemy.orm import Session
+from fastapi import APIRouter
+from starlette import status
 
+from dependencies import DBDependency, UserDependency
+from models import Recipe, RecipeNote
 from schemas import Message, RecipeNoteCreate, RecipeNoteOut, RecipeNoteUpdate
-from models import Recipe, RecipeNote, User
+from utils.http_exceptions import Forbidden, NotFound
+from utils.statements import ensure_exists
+
+router = APIRouter(prefix="/recipe-notes", tags=["RecipeNotes"])
 
 
-router = APIRouter(
-    prefix="/recipe-notes",
-    tags=["RecipeNotes"]
-)
-
-
-@router.post("", response_model=RecipeNoteOut)
-def create_recipe_note(recipe_note: RecipeNoteCreate, db: Session = Depends(get_db)):
-    if db.query(Recipe).filter(Recipe.id == recipe_note.recipe_id).first() is None:
-        raise HTTPException(status_code=404, detail="Recipe not found")
-    
-    if db.query(User).filter(User.id == recipe_note.user_id).first() is None:
-        raise HTTPException(status_code=404, detail="User not found")
+@router.post("", response_model=RecipeNoteOut, status_code=status.HTTP_201_CREATED)
+def create_recipe_note(
+    recipe_note: RecipeNoteCreate, db: DBDependency, user: UserDependency
+):
+    ensure_exists(db, Recipe.id == recipe_note.recipe_id, NotFound("Recipe not found"))
 
     db_recipe_note = RecipeNote(
         recipe_id=recipe_note.recipe_id,
-        user_id=recipe_note.user_id,
-        content=recipe_note.content
+        user_id=user.id,
+        content=recipe_note.content,
     )
 
     db.add(db_recipe_note)
@@ -32,22 +28,31 @@ def create_recipe_note(recipe_note: RecipeNoteCreate, db: Session = Depends(get_
 
     return db_recipe_note
 
+
 @router.get("/note/{note_id}", response_model=RecipeNoteOut)
-def get_recipe_note(note_id: int, db: Session = Depends(get_db)):
+def get_recipe_note(note_id: int, db: DBDependency):
     db_recipe_note = db.query(RecipeNote).filter(RecipeNote.id == note_id).first()
 
     if not db_recipe_note:
-        raise HTTPException(status_code=404, detail="Recipe note not found")
-    
+        raise NotFound("Recipe note not found")
+
     return db_recipe_note
 
+
 @router.patch("/note/{note_id}", response_model=RecipeNoteOut)
-def edit_recipe_note(note_id: int, patch: RecipeNoteUpdate, db: Session = Depends(get_db)):
+def edit_recipe_note(
+    note_id: int, patch: RecipeNoteUpdate, db: DBDependency, user: UserDependency
+):
     db_recipe_note = db.query(RecipeNote).filter(RecipeNote.id == note_id).first()
 
     if not db_recipe_note:
-        raise HTTPException(status_code=404, detail="Recipe note not found")
-    
+        raise NotFound("Recipe note not found")
+
+    if db_recipe_note.user_id != user.id and not user.is_admin:
+        raise Forbidden(
+            "Forbidden: You do not have permissions to edit this recipe note."
+        )
+
     setattr(db_recipe_note, "content", patch.content)
 
     db.commit()
@@ -55,34 +60,43 @@ def edit_recipe_note(note_id: int, patch: RecipeNoteUpdate, db: Session = Depend
 
     return db_recipe_note
 
+
 @router.delete("/note/{note_id}", response_model=Message)
-def delete_recipe_note(note_id: int, db: Session = Depends(get_db)):
+def delete_recipe_note(note_id: int, db: DBDependency, user: UserDependency):
     db_recipe_note = db.query(RecipeNote).filter(RecipeNote.id == note_id).first()
 
     if not db_recipe_note:
-        raise HTTPException(status_code=404, detail="Recipe note not found")
-    
+        raise NotFound("Recipe note not found")
+
+    if db_recipe_note.user_id != user.id and not user.is_admin:
+        raise Forbidden(
+            "Forbidden: You do not have permissions to delete this recipe note."
+        )
+
     db.delete(db_recipe_note)
     db.commit()
 
     return Message(detail="Recipe note deleted successfully")
 
-@router.get("/{recipe_id}", response_model=list[RecipeNoteOut])
-def get_notes_for_recipe(recipe_id: int, db: Session = Depends(get_db)):
-    db_recipe = db.query(Recipe).filter(Recipe.id == recipe_id).first()
 
-    if not db_recipe:
-        raise HTTPException(status_code=404, detail="Recipe not found")
-    
+@router.get("/{recipe_id}", response_model=list[RecipeNoteOut])
+def get_notes_for_recipe(recipe_id: int, db: DBDependency):
+    ensure_exists(db, Recipe.id == recipe_id, NotFound("Recipe not found"))
+
     return db.query(RecipeNote).filter(RecipeNote.recipe_id == recipe_id).all()
 
-@router.delete("/{recipe_id}", response_model=Message)
-def delete_all_notes_from_recipe(recipe_id: int, db: Session = Depends(get_db)):
-    db_recipe = db.query(Recipe).filter(Recipe.id == recipe_id).first()
 
-    if not db_recipe:
-        raise HTTPException(status_code=404, detail="Recipe not found")
-    
+@router.delete("/{recipe_id}", response_model=Message)
+def delete_all_notes_from_recipe(
+    recipe_id: int, db: DBDependency, user: UserDependency
+):
+    ensure_exists(db, Recipe.id == recipe_id, NotFound("Recipe not found"))
+
+    if not user.is_admin:
+        raise Forbidden(
+            "Forbidden: You do not have permissions to delete all recipe notes of this recipe."
+        )
+
     db.query(RecipeNote).filter(RecipeNote.recipe_id == recipe_id).delete()
     db.commit()
